Moves OAuth code into a separate package
